+++ Sources/TuistKit/Services/GenerateService.swift
@@ -4,38 +4,45 @@ import TuistLoader
 import TuistSupport
 
 protocol GenerateServiceProjectGeneratorFactorying {
-    func generator(cache: Bool, includeSources: Set<String>) -> ProjectGenerating
+    func generator() -> ProjectGenerating
 }
 
 final class GenerateServiceProjectGeneratorFactory: GenerateServiceProjectGeneratorFactorying {
-    func generator(cache: Bool, includeSources: Set<String>) -> ProjectGenerating {
-        ProjectGenerator(graphMapperProvider: GraphMapperProvider(cache: cache, includeSources: includeSources))
+    func generator() -> ProjectGenerating {
+        ProjectGenerator(graphMapperProvider: GraphMapperProvider(cache: false, includeSources: Set()))
     }
 }
 
 final class GenerateService {
     // MARK: - Attributes
 
+    private let opener: Opening
     private let clock: Clock
     private let projectGeneratorFactory: GenerateServiceProjectGeneratorFactorying
 
+    // MARK: - Init
+
     init(clock: Clock = WallClock(),
+         opener: Opening = Opener(),
          projectGeneratorFactory: GenerateServiceProjectGeneratorFactorying = GenerateServiceProjectGeneratorFactory())
     {
         self.clock = clock
+        self.opener = opener
         self.projectGeneratorFactory = projectGeneratorFactory
     }
 
     func run(path: String?,
              projectOnly: Bool,
-             cache: Bool,
-             cacheSources: Set<String>) throws
+             open: Bool) throws
     {
         let timer = clock.startTimer()
         let path = self.path(path)
-        let generator = projectGeneratorFactory.generator(cache: cache, includeSources: cacheSources)
+        let generator = projectGeneratorFactory.generator()
 
-        _ = try generator.generate(path: path, projectOnly: projectOnly)
+        let generatedProjectPath = try generator.generate(path: path, projectOnly: projectOnly)
+        if open {
+            try opener.open(path: generatedProjectPath, wait: false)
+        }
 
         let time = String(format: "%.3f", timer.stop())